Version 0.2.0.7
Adaptation for the new architecture. Minor corrections.
Add DataSetsFilter_FastFull for RECO,PU25, PUpmx25 & miniAOD.
A little pbm with refresh : when clicking on an option (rel pmx for example), refresh is not automatic for the opposite option (as ref pmx).
Add getCheckedOptions function to resume all options clicked into QGBoxListsUpdate.
I renamed PU and pmx as PU25 and PUpmx25.