Version 0.3.0.0
Minor correction to have release up to the right window for release and reference choice.

Creation of the function fillQLW_rel1 for filling QLW list. Idem with fillQLW-rel2 & QLW_ref2

New architecture ; the optionGrp is not enabled (greyed) and it is only enabled for tasks_counter == 2. This will be more simple to test all the options.

Fast vs Full is corrected : we can go to Fast vs Full and back to RECO, keeping the old reference and the associated lists.