Version 0.3.0.2
The StdDev buttons are also grayed and they are not for step 4 (resuming).

It will here that you can chose the path where saves are located.

Transfering PART2 (middle part with QtextEdit) into an ovalMiddleGp.py file.

Add a button for showing the list of operations.

Remove QTextEdit in the middle part & replacing it with Label. Note : When changing toward Fast vs Full, the releases are not modified.